feat:add and implement update inactive employee Test

+++ src/main/java/org/example/springboottest/Service/EmployeeService.java
@@ -21,11 +21,13 @@ public class EmployeeService {
 
     public Map<String,Object> addEmployee(Employee employee) throws EmployeeNotQualifiedException, EmployeeAlreadyExistsException {
         int age=employee.getAge();
-        if(age<18||age>65){
-            throw new EmployeeNotAmongLeaglAgeException("Employee age must be between 18 and 65");
-        }
-        if(age>30&&employee.getSalary()<20000){
-            throw new EmployeeNotQualifiedException("For employees older than 30, the salary must be at least 20000");
+        if (age!=0) {
+            if(age<18||age>65){
+                throw new EmployeeNotAmongLeaglAgeException("Employee age must be between 18 and 65");
+            }
+            if(age>30&&employee.getSalary()<20000){
+                throw new EmployeeNotQualifiedException("For employees older than 30, the salary must be at least 20000");
+            }
         }
 
         List<Employee> sameNameGenderEmployees = employeeRepository.getAllEmployeesByGender(employee.getGender());
@@ -60,11 +62,26 @@ public class EmployeeService {
 
     }
 
-    public Employee updateEmployeeById(int id, Employee updatedEmployee) {
+    public Employee updateEmployeeById(int id, Employee updatedEmployee) throws EmployeeAlreadyInactiveException {
+        Employee employee = employeeRepository.getEmployeeById(id);
+        if (employee == null) {
+            throw new EmployeeNotFoundException("Employee with id " + id + " not found");
+        }
+        if (!employee.getStatus()) {
+            throw new EmployeeAlreadyInactiveException("Employee with id " + id + " is inactive");
+        }
         return employeeRepository.updateEmployeeById(id,updatedEmployee);
     }
 
-    public ResponseEntity<Void> deleteEmployeeById(int id) {
+    public ResponseEntity<Void> deleteEmployeeById(int id) throws EmployeeAlreadyInactiveException {
+        Employee employee= employeeRepository.getEmployeeById(id);
+        if(employee==null){
+            throw new EmployeeNotFoundException("Employee with id "+id+" not found");
+        }
+        if(!employee.getStatus()){
+            throw new EmployeeAlreadyInactiveException("Employee with id "+id+" is already inactive");
+        }
+
         return employeeRepository.deleteEmployeeById(id);
     }
